INFINITY-3500 Upstream default service state manager
With this addition, the developer is no longer required to manage the list of active services. Instead, the developer only needs to implement a method which consistently constructs or reconstructs the service from its name and any context data (up to 100KB for the developer to store any info/config for the service).